+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [1.0.6] - 17 April 2021.
+
+* Added `scrollToTopDuration` and `scrollToTopCurve` parameters.
+
+## [1.0.5] - 17 April 2021.
+
+* Alignment fix.
+
+## [1.0.4] - 17 April 2021.
+
+* Removed promptTheme assert.
+
+## [1.0.3] - 17 April 2021.
+
+* Updated description.
+
 ## [1.0.0] - 17 April 2021.
 
 * Initial Release.